refactor: kill testing library
