+++ src/components/App/App.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import logo from './logo.svg';
+import logo from './assets/logo.svg';
 import styles from './App.module.css';
 
 function App() {
